Add roleholder details to GroupMembership
I've decided to make roleholdings happen on the group membership instead of a seperate GourpUser mode. This will make lots of queries more efficient and simpler in the future. It does mean you can't have non-members be a roleholder in a group, but I'm happy with that for now.

This commit just adds the information to GroupMembership. I'll remove the GroupUser models in a bit.